make mt_kahypar_hypernode_id_t a long
We rely on values greater than what a four byte integer can hold to
detect overflows

+++ include/mtkahypartypes.h
@@ -79,8 +79,8 @@ typedef struct {
   mt_kahypar_partition_type_t type;
 } mt_kahypar_partitioned_hypergraph_const_t;
 
-typedef unsigned long int mt_kahypar_hypernode_id_t;
-typedef unsigned long int mt_kahypar_hyperedge_id_t;
+typedef unsigned long long mt_kahypar_hypernode_id_t;
+typedef unsigned long long mt_kahypar_hyperedge_id_t;
 typedef int mt_kahypar_hypernode_weight_t;
 typedef int mt_kahypar_hyperedge_weight_t;
 typedef int mt_kahypar_partition_id_t;
